Remove the pageviewer struct
